Added deprec:rack recipes
The Rack recipes give us a way of introducing a new leaner
set of recipes for deploying Ruby web apps. The idea is that
they wil not have any before or after hooks into callbacks
for the standard Capistrano deploy tasks.

:before and :after hooks will instead live in config for
the project.

This allows us to avoid people being surprised when deprec
is loaded.